Fix undefined behavior in case of time overflow at 32-bit platforms

+++ src/Platforms/armcc/UtestPlatform.cpp
@@ -107,10 +107,10 @@ void (*PlatformSpecificRestoreJumpBuffer)() = PlatformSpecificRestoreJumpBufferI
 *  In Keil MDK-ARM, clock() default implementation used semihosting.
 *  Resolutions is user adjustable (1 ms for now)
 */
-static long TimeInMillisImplementation()
+static unsigned long TimeInMillisImplementation()
 {
     clock_t t = clock();
-    return t;
+    return (unsigned long)t;
 }
 
 ///////////// Time in String
@@ -121,7 +121,7 @@ static const char* DummyTimeStringImplementation()
     return ctime(&tm);
 }
 
-long (*GetPlatformSpecificTimeInMillis)() = TimeInMillisImplementation;
+unsigned long (*GetPlatformSpecificTimeInMillis)() = TimeInMillisImplementation;
 const char* (*GetPlatformSpecificTimeString)() = DummyTimeStringImplementation;
 
 int (*PlatformSpecificVSNprintf)(char *str, size_t size, const char* format, va_list args) = vsnprintf;
